+++ test/integration/test_examples.py
@@ -20,6 +20,8 @@ class TestTutorial:
         "name",
         [
             "custom_backbone",
+            "custom_image_label_loss",
+            "custom_parameterized_image_label_loss",
         ],
     )
     def test_registry(self, name: str):